Add Label component that accepts for attribute
