+++ src/components/form/label/index.js
@@ -1,0 +1,1 @@
+export * from './label'
