+++ src/components/form/label/label.native.js
@@ -1,0 +1,3 @@
+import { Text } from '@keg-hub/keg-components'
+
+export const Label = Text
