+++ src/components/form/label/label.web.js
@@ -1,0 +1,6 @@
+import { unstable_createElement as createElement } from 'react-native-web'
+
+export const Label = props => {
+  const { for: htmlFor, ...rest } = props
+  return createElement('label', { htmlFor, ...rest })
+}
