+++ src/components/tapIndex.js
@@ -1,5 +1,6 @@
 export * from './dates'
 export * from './labels'
+export * from './form'
 export * from './grid'
 export * from './modals'
 export * from './sessionTime'
